combine annotation defaults files & shape defaults files
